Skip distributed state tests when custom is not available
